fix: extract requireAuth to shared module and handle PUT conflict in users API
