+++ src/app/api/users/route.ts
@@ -1,22 +1,7 @@
 import { NextResponse } from 'next/server';
 import { getUsers, createUser } from '@/lib/db/users';
-import { getSetting } from '@/lib/db/settings';
 import { CreateUserSchema } from '@/lib/validators/users';
-
-function requireAuth(): NextResponse | null {
-  const enabled = getSetting('auth_enabled');
-  if (!enabled) {
-    return NextResponse.json(
-      {
-        success: false,
-        error: 'User management requires auth to be enabled',
-        code: 'AUTH_NOT_ENABLED',
-      },
-      { status: 403 }
-    );
-  }
-  return null;
-}
+import { requireAuth } from '@/lib/auth-guard';
 
 export async function GET() {
   try {
